requirements.txt
# Phase 1 — Ingestion
sentence-transformers==3.3.1
pinecone==6.0.2
pymupdf==1.25.5
python-docx==1.1.2
langchain==0.3.25
langchain-community==0.3.23

# Phase 3 — API
fastapi==0.115.0
uvicorn==0.34.0
pydantic==2.11.3

# Phase 4 — UI
streamlit==1.45.1

# Utilities
python-dotenv==1.1.0
pyyaml==6.0.2
tiktoken==0.9.0
